fix: optimize brush painting
* ref: fix complier warnings

* ref: add compile flags for CMakeLists.txt

* ref: separate windows file into dirs 'Fonts' 'Images'

* ref: new #define = sf::KeyBoard is pressed

* ref: use shared_ptr instead of static in func LoadButtonImages

* ref: optimized LoadButtonImg

* fix: brush radius is int now

* fix: brush history is working now

+++ App/Brush/Brush.hpp
@@ -21,7 +21,7 @@ class BrushColorDisplay;
 class Brush {
  private:
     sf::CircleShape shape_;
-    float radius_;
+    int radius_;
     sf::Color color_;
 
     sf::Texture brushCursorTexture;
